Added start to Timer/Counter 0 runtime
extra files not necessarily to be included or compiled but for scratch
work.  runtime_count source and header files.  lan_v3_sketch has the
more final form for the code, but is not fully tested.  newest code
runs but doesn’t have all the capabilities of running a charging,
lighting, runtime-counting lantern.  Need to add macros for setup of
TC0!